Доработал функции, разбил файл db на логические блоки
