+++ database/connection.php
@@ -1,0 +1,6 @@
+<?php
+
+$connect = mysqli_connect("localhost", "root", "root", "readme");
+if ($connect === false) {
+    die("Ошибка подключения: " . mysqli_connect_error());
+}
