+++ functions.php
@@ -267,6 +267,17 @@ function getFilePath($full_path, $file_name): string {
     return $full_path . basename($file_name);
 }
 
+function downloadPictureFile($file_path, $uploads_dir, $file_name, $temp_path, $errors): string {
+    $result = "";
+
+    if (count($errors) === 0 &&
+        move_uploaded_file($temp_path, $file_path)) {
+        $result = $uploads_dir . $file_name;
+    }
+
+    return $result;
+}
+
 function downloadContent($file_path, $uploads_dir, $file_name, $data, $errors): string {
     $result = "";
 
@@ -303,6 +314,83 @@ function checkYoutubeURL($url): string {
         $result = "Неверная ссылка на видео";
     } else {
         $result = check_youtube_url($url);
+    }
+
+    return $result;
+}
+
+function addPictureFile($web_name, $result, $full_path, $uploads_dir): array {
+    if ($_FILES[$web_name]["error"] === 0) {
+        $picture = $_FILES[$web_name];
+        $result["errors"] = addError(
+            $result["errors"],
+            checkPictureType("/image\/(jpg|jpeg|png|gif)/i", $picture["type"]),
+            $web_name
+        );
+        $file_path = getFilePath($full_path, $picture["name"]);
+        $result["content"] = downloadPictureFile(
+            $file_path,
+            $uploads_dir,
+            $picture["name"],
+            $picture["tmp_name"],
+            $result["errors"]
+        );
+    }
+
+    return $result;
+}
+
+function addPictureURL($web_name, $result, $field, $full_path, $uploads_dir): array {
+    if ($result["content"] === "") {
+        if ($_POST[$web_name] === "") {
+            $result["errors"] = addError($result["errors"], "Необходимо выбрать изображение с компьютера или указать ссылку из интернета.", $web_name);
+            return $result;
+        }
+
+        $picture_url = filter_var($_POST[$web_name], FILTER_VALIDATE_URL);
+        $photo_info = pathinfo($picture_url);
+        $result["errors"] = addError($result["errors"], checkPictureType("/(jpg|jpeg|png|gif)/i", $photo_info["extension"] ?? ""), $web_name);
+
+        if (count($result["errors"]) === 0) {
+            $download_photo = file_get_contents($picture_url);
+            $file_path = getFilePath($full_path, $photo_info["basename"]);
+            $result["content"] = downloadContent($file_path, $uploads_dir, $photo_info["basename"], $download_photo, $result["errors"]);
+            $result[$field] = $picture_url;
+        }
+    }
+
+    return $result;
+}
+
+function addWebsite($web_name, $result, $field): array {
+    $website = $_POST[$web_name];
+    $result["errors"] = addError($result["errors"], checkURL($website), $web_name);
+    $result["content"] = getValidateURL($website, $result["errors"]);
+    $result[$field] = $result["content"];
+
+    return $result;
+}
+
+function addVideoURL($web_name, $result, $field): array {
+    $video_url = $_POST[$web_name];
+    $result["errors"] = addError($result["errors"], checkYoutubeURL($video_url), $web_name);
+    $result["content"] = getValidateURL($video_url, $result["errors"]);
+    $result[$field] = $result["content"];
+
+    return $result;
+}
+
+function addTextContent($web_name, $result, $field, $required_empty_filed): array {
+    $result["errors"] = addError($result["errors"], checkFilling($web_name, $required_empty_filed[$web_name]), $web_name);
+    $result[$field] = $_POST[$web_name] ?? "";
+
+    return $result;
+}
+
+function addTags($field, $result): array {
+    if (isset($_POST[$field]) && $_POST[$field] !== "") {
+        $result[$field] = explode(" ", $_POST[$field]);
+        $result["errors"] = checkTags("/^#[A-Za-zА-Яа-яËё0-9]{1,19}$/", $result[$field], $result["errors"]);
     }
 
     return $result;
